test(resourcemanagers): centralize kafka and ping source fixtures
Shared event-source constants and entry builders live in fixtures
so ping tests no longer depend on kafka-only definitions. Ping
setup reuses the shared entry pattern instead of repeated inline
config structs.
